requirements.txt
fastmcp>=2.0.0
anthropic>=0.40.0
pydantic>=2.0.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
